rename top matrix workflow object 'swift'->'config'
Motivation:

The use of 'swift' here is very arbitrary and adds no value. The object
is only there at all because GitHub Actions seems to require it.

Modifications:

Rename the top-level object from 'swift' to 'config' to maybe add value
to the intent of the object and hopefully be less cryptic.

Result:

Everything should just keep working.