Strict concurrency for SelectableEventLoop & MTELG
Motivation

SelectableEventLoop and MTEL are our major entry points for
execution. These need to be strict concurrency clean. Unfortunately,
changes in one tend to ripple into the other, so we need to tackle
both at once.

Ideally we would not use `@unchecked Sendable` for
SelectableEventLoop. With NIOLoopBound and our other tools, I
think we have a real shot at getting the compiler to prove much
more of the correctness of this type. However, right now that's
not prudent: we'd like to unblock the rest of the ecosystem. So
we'll tolerate an `@unchecked Sendable` for now, with a plan to
revisit.

Modifications

- SelectableEventLoop is @unchecked Sendable
- A few closures get annotated @Sendable
- Add a NIOLockedValueBox to convince the compiler that this use
    is safe.
- Make a few types explicitly not Sendable

Result

No concurrency warnings in MTELG and SelectableEventLoop.