Avoid using stderr explicitly
Motivation

It's tempting for us to use the FILE * for stderr whenever we
want to print to it. However, the Linux libc modules aren't
appropriately annotated to call that field constant, and indeed it
isn't. This causes Swift to throw a fit if we even touch it.

Modifications

Introduce an awkward function that issues direct writes to
stderr instead of using the FILE *.

Result

Swift is happier.