unchecked Sendable for the Channels (#3050)
### Motivation:

All of NIO's channels are thread-safe, and that includes the NIOPosix
ones. In an ideal world we'd have written these in such a way that they
rely on very small, easily-verified types to enforce this constraint. We
will migrate towards that in time, but for expediency we can mark these
as they are, which is `@unchecked Sendable`.

### Modifications:

Mark the core NIOPosix channel types `@unchecked Sendable`.

### Result:

Fewer warnings.