Test allocations of isolated ELF views
Motivation:

We've added isolated views onto EventLoopFutures. These are great,
but as initially implemented they add a bunch of overhead to the nonisolated
versions. That's not optimal, but before we change the code we should make sure
we have a benchmark baseline.

Modifications:

- Extend the test_future_lots_of_callbacks alloc test to add all the callbacks
    that the isolated views have.
- Duplicate that test to one that adds isolated wrappers so we can compare.

Result:

Two regression tests whose allocation counts should be equal, but aren't.