Add allocation counter tests for isolated EL operations
Motivation:

As with the ELF operations before them, the isolated EL operations currently
incur overhead above-and-beyond the overhead of their non-isolated counterparts.
That's not what we want to see. However, before we "fix" them, we need to add
regression testing to confirm our fix actually worked.

Modifications:

- Add isolated variations of all current EL scheduling tests.
- Where isolated methods didn't have alloc counter tests, add new ones
    for the non-isolated versions too.

Result:

More tests.