Move Isolated EL operations to EL protocol
Motivation

Right now the isolated view on the EventLoop has to implement
its interface in terms of UnsafeTransfer, because there are no
non-Sendable operations it can use.

That hurts us because we have to allocate an extra closure for all
these operations, making them slower than they need to be. This
is unavoidable in some contexts, but most EventLoops would be able
to offer a fast-path that avoids this extra allocation.

To achieve that, we need to move this logic into a customisation
point on the EventLoop protocol. Our fallback default implementation
can be the same as we do today.

Modifications

- Add customisation points for EventLoop for each of its operations
    that has a protocol witness already.
- Add default implementations of these, using the implementation
    from EventLoop.Isolated.

Results

No behavioural changes, but the code has moved.