Improve coverage of isolated views
Motivation:

We were only testing the isolated views on SelectablEventLoop.
That was fine when all the implementation was common, but it isn't going to
be anymore, so we should start having tests across all our EL kinds.

Modifications:

Refactor the tests to take their EL as an argument
Add support for ELs that need to be explicitly run
Add tests for NIOAsyncTestingEL and EmbeddedEL

Result:

Better test coverage.