Ensure we test the isolated fallback methods
Motivation

I'm about to change all our main event loops to use our new customisation
points. That's great, but it'll have the effect of removing all the testing
of our default implementations, which is not a good idea. So let's add
some more tests.

Modifications

Define a "fallback" event loop for testing the isolated
methods.

Result

Reduced likelihood of regressions in future.