Provide the remaining buffer when the NIOAsyncSequenceProducer terminates
# Motivation

Our `NIOAsyncChannel` abstractions is running into crashes in the bootstraps and multiplexer when a `NIOAsyncChannel` is buffered inside an async stream and the consuming task gets cancelled.

# Modification

This PR modifies the `NIOAsyncSequenceProducer` to hand over the remaining elements once it is finished to `didTerminate` to allow the user to run any manual clean up such as closing the underling channel of a `NIOAsyncChannel`.

# Result

A path forward to avoid the crashes we are seeing.