Make sure NIOAsyncWriter test doesn't hang indefinitely
The testSuspendingBufferedYield_whenWriterFinished test causes the test
application to hang in the Android emulator. This fix sets a timeout on
waiting for the ConditionLock, so we don't wait indefinitely.

Additionally, the timeout for waiting for both yields being suspended is
increased to make it less likely that an incorrect state is reached.