Update re-usable workflows to use nightly_next (#3122)
### Motivation:

Updating the CI parameters for changing nightly job targets with each
new Swift version (6.0, 6.1, ...) is onerous. Whilst this gives us some
value in the non-nightly pipelines by forcing adopters to apply thought
to what is applied to each supported version, the nightlies arguably
deliver less value in this regard as they tend to be non-required.

### Modifications:

Change parameters to refer to nightly_next rather than nightly_6_1,
implement
a shim so that the old parameter names still work.

### Result:

We will be able to modify top-level workflows to use nightly_next
terminology.

Test run with an additional commit opting in to use these changes at the
top level
https://github.com/apple/swift-nio/actions/runs/13524774485/job/37792437713?pr=3122