Enable macOS CI
Motivation:

To expand test coverage to the Darwin platforms.

Modifications:

Create a new re-usable workflow 'macos_tests.yml' which offers testing
and compilation checks for multiple Xcodes and simulator destinations.

The new workflow first performs `swift test` with configurable
parameters and then goes on to build the code targetting generic targets
for:
* macOS
* iOS
* watchOS
* tvOS
* visionOS

At the moment this is workflow is enabled to run on the NIO repository:
* on each commit to a PR
* on each merge to `main`
* overnight on a timer

This will likely be revised due to capacity.

Result:

Increased test coverage.