A re-usable CMake check workflow (#3144)
### Motivation

A few repositories have custom workflows for checking CMake, we would
like to unify this for maintenance and adoption simplicity.

Further to this at the moment they only check that the scripts are
up-to-date, we'd also like to have them build using CMake.

### Modifications

Add the re-usable workflow and supporting scripts

### Result:

Repositories using CMake can switch to the centralised workflow.

An example of it in action
https://github.com/apple/swift-crypto/actions/runs/13837092514/job/38782284575