Speed up allocation counting tests
Motivation:

The allocation counting tests take around 18 minutes in CI, which is
quite a while! The tests themselves don't take nearly so long, however
each test is run 11 times but between each time the framework waits for
allocations to become rebalanced (i.e. everything we allocated should
be freed). The framework will wait up to 5 seconds between each
iteration for this. About 15 or so tests take 55 seconds, meaning
they're just waiting for the allocs to quiece. It turns out that
dispatch doesn't always free blocks from `async` work items immediately.
This is relevant because a number of tests create their own MTELGs, and
so when they are shutdown we see allocations from the shutdown flow
lingering for more than 5 seconds.

Not only does this slow the tests down, if the allocations don't quiece,
the framework ignores the result, so we lose some coverage too.

Modifiications:

- Add a pre-warmed singleton MTELG which tests can use, this ensures it
  has plenty of space in its queues so that intermediate allocs from
  running tests aren't seen as leaks
- Cancel tasks which are scheduled to run after the test to avoid
  'leaks'.
- Avoid using `swift run` and just execute the test binary, this shaves
  off a few seconds for each test.
- Print out how long each test run takes to make debugging easier

Result:

- Faster alloc tests