Execute swift --version in shared workflows (#3247)
### Motivation:

To have more version about which swift is running to aid in debugging.

### Modifications:

Run swift --version on common workflows, this should be safe because
they all assume Swift is already nistalled unlike the Static SDK
workflow.

### Result:

Swift tools information will be printed before job execution

An example of this working:
https://github.com/apple/swift-nio/actions/runs/15049159516/job/42299334292