Adjust for SendableMetatype
Motivation

With the introduction of isolated conformances, it has become
necessary to start managing the use of metatypes for some of
our protocols. In general, we don't want to force the relevant
protocols to only be conformed in non-isolated forms. Instead,
we just want to make the specific APIs non-usable.

Modifications

- Add shims for SendableMetatype that only use it when it is
    available.
- Require SendableMetatype where needed, gated by @preconcurrency.

Result

We continue to be safe.