Extend NIOEventLoopTickInfo with end-time and loop sleep time
Motivation:

NIO’s event loop can emit metrics. Currently, this includes the start time (i.e. when the loop started working) and the number of tasks which were executed in that tick.

It would be useful to collect the time the loop tick ended and the length of time the loop was asleep before starting the tick.

Modifications:

The event loop has a new property to track the time between runs. The tick info collects two new values:
* The time since the last run.
* The time the loop finished.

Result:

The new metrics are available for collection.