Work around Android nullability errors
Motivation:

Android messed up their nullability annotations in fts_open.
The original change in https://android.googlesource.com/platform/bionic/+/dec8efd72a6ad8b807a15a614ae1519487cfa456
asserted that fts_open's path argument took a non-null pointer to
an array of non-null pointers to strings. That's challenging,
because the _end_ of that array is indicated by a null pointer.

Regardless, this was eventually fixed in
https://android.googlesource.com/platform/bionic/+/da81ec4d1cbd0279014feb60535bf38defcd9346.
Unfortunately, that was more than a year after the offending change,
so we need to work around it.

Modifications:

Add a shim for Android that omits the nullability annotations.
Use that shim on Android.

Result:

Android works again.
Resolves #3273.